+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## [Unreleased]
 
+- Add `Program#attach_raw_tracepoint`
 - Add `Program#attach_tracepoint`
 - Add `Program#attach_uprobe`
 - Add `Program#attach_kprobe`